requirements.txt
python-dotenv
openai
rich
tiktoken
langchain
langgraph
langchain-community
langchain-openai
beautifulsoup4
requests
pypdf